fix: all of the jwt system

+++ src/main/java/org/example/memoaserver/global/service/RefreshTokenService.java
@@ -1,4 +1,4 @@
-package org.example.memoaserver.global.security.jwt.service;
+package org.example.memoaserver.global.service;
 
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
@@ -22,7 +22,7 @@ import java.util.concurrent.TimeUnit;
 public class RefreshTokenService {
     private final RedisTemplate<String, Object> redisTemplate;
     private static final String TOKEN_PREFIX = "refresh_token:";
-    private static final String INVERSE_INDEX_PREFIX = "refresh_to_username:";
+    private static final String INVERSE_INDEX_PREFIX = "refresh_to_email:";
     private final JwtUtil jwtUtil;
     private final JwtProperties jwtProperties;
 
@@ -42,11 +42,11 @@ public class RefreshTokenService {
         }
 
         if (refresh == null) {
-            return new ResponseEntity<>("refresh cookie not found", HttpStatus.UNAUTHORIZED);
+            return new ResponseEntity<>("refresh token null", HttpStatus.BAD_REQUEST);
         }
 
         try {
-            jwtUtil.getUsername(refresh);
+            jwtUtil.getEmail(refresh);
         } catch (Exception e) {
             return new ResponseEntity<>("unknown token", HttpStatus.UNAUTHORIZED);
         }
@@ -60,23 +60,28 @@ public class RefreshTokenService {
 
         String category = jwtUtil.getCategory(refresh);
 
-        if (!existsByRefreshToken(refresh)) {
-            return new ResponseEntity<>("refresh token not found", HttpStatus.NOT_FOUND);
+        if (!category.equals("refresh")) {
+
+            return new ResponseEntity<>("invalid refresh token", HttpStatus.BAD_REQUEST);
         }
 
-        String username = jwtUtil.getUsername(refresh);
+        if (!existsByRefreshToken(refresh)) {
+            return new ResponseEntity<>("refresh token not found", HttpStatus.BAD_REQUEST);
+        }
+
+        String email = jwtUtil.getEmail(refresh);
         String role = jwtUtil.getRole(refresh);
 
-        String newAccess = jwtUtil.createJWT("access", username, role, jwtProperties.getAccess().getExpiration());
-        String newRefresh = jwtUtil.createJWT("refresh", username, role, jwtProperties.getRefresh().getExpiration());
+        String newAccess = jwtUtil.createJwt("access", email, role, jwtProperties.getAccess().getExpiration());
+        String newRefresh = jwtUtil.createJwt("refresh", email, role, jwtProperties.getRefresh().getExpiration());
 
         deleteByRefreshToken(refresh);
-        addRefreshEntity(username, newRefresh, jwtProperties.getRefresh().getExpiration());
+        addRefreshEntity(email, newRefresh, jwtProperties.getRefresh().getExpiration());
 
         response.setHeader("Authorization", newAccess);
-        response.addCookie(createCookie("refresh", newRefresh, jwtProperties.getRefresh().getExpiration()));
+        response.addCookie(createCookie("refresh", newRefresh,jwtProperties.getRefresh().getExpiration()));
 
-        return new ResponseEntity<>("success", HttpStatus.OK);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     public void logout(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws IOException, ServletException {
@@ -103,8 +108,8 @@ public class RefreshTokenService {
         try {
             jwtUtil.isExpired(refresh);
         } catch (ExpiredJwtException e) {
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            return;
+
+            response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
         }
 
         String category = jwtUtil.getCategory(refresh);
@@ -128,37 +133,27 @@ public class RefreshTokenService {
         response.setStatus(HttpServletResponse.SC_OK);
     }
 
-    public Cookie createCookie(String key, String value, Long maxAge) {
-        Cookie cookie = new Cookie(key, value);
-        int maxAgeInt = maxAge.intValue();
-        cookie.setMaxAge(maxAgeInt);
-        // cookie.setSecure(true);
-        // cookie.setPath("/");
-        cookie.setHttpOnly(true);
-        return cookie;
+    public void deleteByRefreshToken(String refresh) {
+        String email = findByRefreshToken(refresh.replaceFirst(INVERSE_INDEX_PREFIX, ""));
+        redisTemplate.delete(TOKEN_PREFIX + email);
+        redisTemplate.delete(INVERSE_INDEX_PREFIX + refresh.replaceFirst(INVERSE_INDEX_PREFIX, ""));
     }
 
-    public String findByPhoneNumber(String username) {
-        String key = TOKEN_PREFIX + username;
-        return (String) redisTemplate.opsForValue().get(key);
-    }
-
-    public void addRefreshEntity(String username, String refresh, Long expiredMs) {
-        String key = TOKEN_PREFIX + username;
+    public void addRefreshEntity(String phoneNumber, String refresh, Long expiredMs) {
+        String key = TOKEN_PREFIX + phoneNumber;
         String refreshKey = INVERSE_INDEX_PREFIX + refresh;
 
         if (Boolean.TRUE.equals(redisTemplate.hasKey(key))) {
-            deleteByRefreshToken(INVERSE_INDEX_PREFIX + findByPhoneNumber(username));
+            deleteByRefreshToken(INVERSE_INDEX_PREFIX + findByEmail(phoneNumber));
         }
 
         redisTemplate.opsForValue().set(key, refresh, expiredMs, TimeUnit.MILLISECONDS);
-        redisTemplate.opsForValue().set(refreshKey, username, expiredMs, TimeUnit.MILLISECONDS);
+        redisTemplate.opsForValue().set(refreshKey, phoneNumber, expiredMs, TimeUnit.MILLISECONDS);
     }
 
-    public void deleteByRefreshToken(String refresh) {
-        String username = findByRefreshToken(refresh.replaceFirst(INVERSE_INDEX_PREFIX, ""));
-        redisTemplate.delete(TOKEN_PREFIX + username);
-        redisTemplate.delete(INVERSE_INDEX_PREFIX + refresh.replaceFirst(INVERSE_INDEX_PREFIX, ""));
+    public String findByEmail(String email) {
+        String key = TOKEN_PREFIX + email;
+        return (String) redisTemplate.opsForValue().get(key);
     }
 
     public String findByRefreshToken(String refresh) {
@@ -170,4 +165,15 @@ public class RefreshTokenService {
         String key = INVERSE_INDEX_PREFIX + refresh;
         return redisTemplate.hasKey(key);
     }
+
+    public Cookie createCookie(String key, String value, Long maxAge) {
+        Cookie cookie = new Cookie(key, value);
+        int maxAgeInt = maxAge.intValue();
+        cookie.setMaxAge(maxAgeInt);
+//        cookie.setSecure(true);
+//        cookie.setPath("/");
+        cookie.setHttpOnly(true);
+        return cookie;
+    }
 }
+
